Creating ClientApplication.ts subclass
